+++ sigils/core.nim
@@ -11,7 +11,7 @@ when defined(sigilsDebug):
 export signals, slots, agents
 
 method callMethod*(
-    ctx: Agent, req: SigilRequest, slot: AgentProc
+    ctx: Agent, req: sink SigilRequest, slot: AgentProc
 ): SigilResponse {.base, gcsafe, effectsOf: slot.} =
   ## Route a sigil request.
   debugPrint "callMethod: normal: ",
@@ -31,12 +31,12 @@ method callMethod*(
 
 when not sigilsSlotEnvDisabled:
   method callMethod*(
-      ctx: Agent, req: SigilRequest, subscription: Subscription
+      ctx: Agent, req: sink SigilRequest, subscription: Subscription
   ): SigilResponse {.base, gcsafe.} =
     ## Route a sigil request through a static slot or an env-backed closure slot.
     if subscription.envSlot.isNil:
       {.cast(gcsafe).}:
-        result = ctx.callMethod(req, subscription.packedSlot)
+        result = ctx.callMethod(ensureMove(req), subscription.packedSlot)
     else:
       {.cast(gcsafe).}:
         subscription.envSlot(ctx, req.params, subscription.env)
@@ -53,14 +53,19 @@ template checkSlotResponse(res: SigilResponse) =
   elif sigilsCborSerdeEnabled:
     discard
   else:
-    variantMatch case res.result.buf as u
+    variantMatch case res.result.payload as u
     of SigilError:
       raise newException(AgentSlotError, $u.code & " msg: " & u.msg)
     else:
       discard
 
 template callSlotsImpl(obj: Agent, req: SigilRequest, subsIter: untyped) =
+  var remainingSubscriptions = 0
+  for _ in subsIter:
+    remainingSubscriptions.inc()
+
   for sub in subsIter:
+    remainingSubscriptions.dec()
     {.cast(gcsafe).}:
       when defined(sigilsDebug):
         if sub.tgt[].freedByThread != 0:
@@ -71,10 +76,17 @@ template callSlotsImpl(obj: Agent, req: SigilRequest, subsIter: untyped) =
           debugPrint "exec:call:obj:id: ", $obj.getSigilId()
           discard c_raise(11.cint)
         assert sub.tgt[].freedByThread == 0
+      var subReq =
+        if remainingSubscriptions == 0:
+          move(req)
+        else:
+          req.clone()
       when sigilsSlotEnvDisabled:
-        var res: SigilResponse = sub.tgt[].callMethod(req, sub.packedSlot)
+        var res: SigilResponse = sub.tgt[].callMethod(
+          ensureMove(subReq), sub.packedSlot
+        )
       else:
-        var res: SigilResponse = sub.tgt[].callMethod(req, sub)
+        var res: SigilResponse = sub.tgt[].callMethod(ensureMove(subReq), sub)
 
       checkSlotResponse(res)
 
@@ -85,35 +97,51 @@ template callSlotsLocalImpl(
     args: untyped,
     subsIter: untyped
 ) =
-  var
-    reqReady = false
-    req: SigilRequest
+  var remainingSubscriptions = 0
+  for _ in subsIter:
+    remainingSubscriptions.inc()
+
   for sub in subsIter:
+    remainingSubscriptions.dec()
     {.cast(gcsafe).}:
       if not sub.directSlot.isNil:
         sub.directSlot(sub.tgt[], addr args)
       else:
-        if not reqReady:
-          req = initSigilRequest[typeof(obj), typeof(args)](
-            procName = procName, args = args, origin = origin
-          )
-          reqReady = true
+        var req =
+          if remainingSubscriptions == 0:
+            initSigilRequest[typeof(obj), typeof(args)](
+              procName = procName,
+              args = move(args),
+              origin = origin,
+            )
+          else:
+            initSigilRequest[typeof(obj), typeof(args)](
+              procName = procName,
+              args = args.clone(),
+              origin = origin,
+            )
         when sigilsSlotEnvDisabled:
-          var res: SigilResponse = sub.tgt[].callMethod(req, sub.packedSlot)
+          var res: SigilResponse = sub.tgt[].callMethod(
+            ensureMove(req), sub.packedSlot
+          )
         else:
-          var res: SigilResponse = sub.tgt[].callMethod(req, sub)
+          var res: SigilResponse = sub.tgt[].callMethod(ensureMove(req), sub)
         checkSlotResponse(res)
 
-method callSlots*(obj: Agent, req: SigilRequest) {.base, gcsafe.} =
-  callSlotsImpl(obj, req, obj.getSubscriptions(req.procName))
+method callSlots*(obj: Agent, req: sink SigilRequest) {.base, gcsafe.} =
+  let procName = req.procName
+  var ownedReq = ensureMove(req)
+  callSlotsImpl(obj, ownedReq, obj.getSubscriptions(procName))
 
-method callSlots*(obj: AgentActor, req: SigilRequest) {.gcsafe.} =
+method callSlots*(obj: AgentActor, req: sink SigilRequest) {.gcsafe.} =
   obj.ensureActorReady()
+  let procName = req.procName
   var subs: seq[Subscription]
   withLock obj.lock:
-    for sub in obj.getSubscriptions(req.procName):
+    for sub in obj.getSubscriptions(procName):
       subs.add(sub)
-  callSlotsImpl(Agent(obj), req, subs.items)
+  var ownedReq = ensureMove(req)
+  callSlotsImpl(Agent(obj), ownedReq, subs.items)
 
 proc callSlotsLocal*[A](
     obj: Agent, procName: SigilName, origin: SigilId, args: var A
@@ -130,11 +158,11 @@ proc callSlotsLocal*[A](
     callSlotsLocalImpl(obj, procName, origin, args, obj.getSubscriptions(procName))
 
 proc emit*(call: (Agent | WeakRef[Agent], SigilRequest)) =
-  let (obj, req) = call
+  var (obj, req) = call
   when obj is WeakRef[Agent]:
-    obj[].callSlots(req)
+    obj[].callSlots(ensureMove(req))
   else:
-    obj.callSlots(req)
+    obj.callSlots(ensureMove(req))
 
 proc emit*[T: Agent, A](call: sink SigilLocalCall[T, A]) =
   var localCall = call